pyproject.toml
[build-system]
requires = ["setuptools>=78", "wheel>=0.41"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pre-commit",
    { include-group = "type-check" },
    { include-group = "lint" },
    { include-group = "test" },
]
type-check = ["mypy", "types-pyaudio"]
lint = ["ruff", "pylint"]
test = ["pytest", "pytest-cov", "pytest-xdist", "pytest-unused-fixtures"]

[project]
name = "baji_recognition"
description = "Identify whether regular or low Baji is speaking"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "librosa>=0.11.0",
    "numpy>=2.4.3",
    "pyannote-audio>=4.0.4",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.2",
]
dynamic= ["version"]

[project.license]
text="TBD"

# Uncomment to add script entry points
[project.scripts]
detect_speaker = "baji_recognition.main:main"

# [tools.setuptools.package-data]
# baji_recognition = ["some_files"]

[tool.setuptools.dynamic]
version = { attr = "baji_recognition.__version__"}

[tool.bandit]
assert_used = { skips = ["**/test_*.py"] }
exclude = ["build/"]

[tool.pylint.messages_control]
disable = [
    "line-too-long",
    "missing-function-docstring",
    "missing-class-docstring",
]

[tool.ruff]
target-version = "py312"
src = ["src"]
line-length = 115

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN",      # I don't like these
    "COM812",   # recommended to ignore by formatter
    "ISC001",   # recommended to ignore by formatter
    "E501",     # ruff format handles line length
    "D100",     # Don't plan on including module-level docstrings
    "D102",     # Don't plan on including module-level docstrings
    "D103",     # Don't plan on including module-level docstrings
    "D203",     # Incompatible with D211 and ruff format
    "D213",     # Incompatible with D212
    "D400",     # I don't care for this docstring formatting nitpick
    "D415",     # I don't care for this docstring formatting nitpick
    "FBT002",    # Disagree
    "T201",    # Disagree
]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    "S101", # asserts allowed in tests...
    "ARG", # Unused function args -> fixtures nevertheless are functionally relevant...
    "FBT", # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
    "PLR2004", # Magic value used in comparison, ...
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
    "D103", # I don't like docstrings *that* much
    "INP001", # tests don't need to be a module
]
[tool.ruff.lint.pylint]
max-args = 6

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 115
disable = """
    line-too-long,
    missing-module-docstring,
"""
max-args = 6

[tool.mypy]
files = ["."]
exclude = ["build/"]
strict = true
explicit_package_bases = true
mypy_path = ["src"]

# add modules that don't have stubs (but those can be added for some imports)
[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true